A Personal Touch
I tried adding a little more to the textures to make a contrast between
the crater and the surrounding area (I also elevated the terrain
immediately next to the crater to make it look like the town had been
ripped out of the ground). Aside from that, this is the same as the
last one. Feel free to disregard this one if you don’t like the changes
BTW, I just applied a texture to the walls rather than replace them
with rocks to give the appearance of sentient life clawing the ground
out rather than make it look like a meteor created it